fix: Fix up cloud view
